// src/types/errors.ts
export enum ErrorCodes {
    INVALID_EMAIL = 'INVALID_EMAIL',
    INVALID_PASSWORD = 'INVALID_PASSWORD',
    INVALID_EMAIL_OR_PASSWORD = 'INVALID_EMAIL_OR_PASSWORD',
    INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
    INVALID_DATA = 'INVALID_DATA',
    INVALID_DOCUMENT = 'INVALID_DOCUMENT',
    INVALID_NAMES = 'INVALID_NAMES',
    INVALID_SURNAMES = 'INVALID_SURNAMES',
    INVALID_PHONE = 'INVALID_PHONE',
    INVALID_TERMS = 'INVALID_TERMS',
    INVALID_TOKEN = 'INVALID_TOKEN',
}